refactor(WIP): start building models module
