+++ cargo-shuttle/src/client.rs
@@ -11,8 +11,9 @@ use reqwest_retry::policies::ExponentialBackoff;
 use reqwest_retry::RetryTransientMiddleware;
 use serde::de::DeserializeOwned;
 use serde::Deserialize;
-use shuttle_common::project::{self, ProjectName};
-use shuttle_common::{deployment, log, secret, service, user, ApiKey, ApiUrl};
+use shuttle_common::models::{deployment, project, user};
+use shuttle_common::project::ProjectName;
+use shuttle_common::{log, secret, service, ApiKey, ApiUrl};
 use tokio::net::TcpStream;
 use tokio_tungstenite::tungstenite::client::IntoClientRequest;
 use tokio_tungstenite::{connect_async, MaybeTlsStream, WebSocketStream};
