+++ cargo-shuttle/src/logger.rs
@@ -1,0 +1,1 @@
+
